+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Yii Framework 2 redis extension Change Log
 
 2.0.13 under development
 ------------------------
-
+- Redis 5.0 stream commands added. Read more at [streams intro](https://redis.io/topics/streams-intro).
 - no changes in this release.